moved course note to get_activity ns
