+++ src/transformer/utils/get_activity/course_note.php
@@ -19,10 +19,12 @@
  *
  * @package   logstore_xapi
  * @copyright Daniel Bell <daniel@yetanalytics.com>
+ *            Milt Reder <milt@yetanalytics.com>
  * @license   https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace src\transformer\utils;
+namespace src\transformer\utils\get_activity;
+use src\transformer\utils as utils;
 
 /**
  * Transformer utility for generating note object for note_created and note_updated events
@@ -33,17 +35,17 @@ namespace src\transformer\utils;
  * @param array $note
  * @return object
  */
-function note_object($config, $lang, $subject, $note) {
+function course_note($config, $lang, $subject, $note) {
   return [
     'id' => $config['app_url'].'/notes/view.php?id='.$note->id,
     'definition' => [
-      'name' => [$lang => get_string_html_removed($note->subject)],
+      'name' => [$lang => utils\get_string_html_removed($note->subject)],
       'type' =>  'http://activitystrea.ms/note',
-      'description' => [$lang => get_string_html_removed($note->content)],
+      'description' => [$lang => utils\get_string_html_removed($note->content)],
       'extensions' => [
         "https://xapi.edlm/profiles/edlm-lms/concepts/activity-extensions/note-type" => "course",
         "https://xapi.edlm/profiles/edlm-lms/concepts/activity-extensions/note-subject" =>
-          get_user($config,$subject)
+          utils\get_user($config,$subject)
       ]
     ]
   ];
